requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SocketIO==5.1.1
gunicorn==20.1.0
gevent==21.12.0
gevent-websocket==0.10.1
celery==5.2.3
redis==4.1.0
webvtt-py==0.4.6
numpy==1.21.6
pandas==1.3.5
openpyxl==3.0.9
python-dotenv==0.19.2
requests==2.27.1
anthropic==0.8.1 